pyproject.toml
[project]
name = "titanicmlcicd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "joblib>=1.5.3",
    "matplotlib>=3.10.0",
    "optuna>=4.7.0",
    "pandas==2.3.3",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pytest>=9.0.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "setuptools==68.2.2",
    "uvicorn>=0.41.0",
    "xgboost>=3.2.0",
    "ydata-profiling==4.18.1",
]
